:bug: fix proto
fix proto

+++ ex01/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: sichoi <sichoi@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/27 16:53:05 by sichoi            #+#    #+#             */
-/*   Updated: 2022/07/28 14:01:38 by sichoi           ###   ########.fr       */
+/*   Updated: 2022/07/28 15:03:26 by sichoi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@
 # include <iostream>
 
 template <typename T>
-void	iter(T *p, unsigned long len, void (*f)(const T& a))
+void	iter(T *p, unsigned int len, void (*f)(const T& a))
 {
-	for (unsigned long i = 0; i < len; ++i)
+	for (unsigned int i = 0; i < len; ++i)
 	{
 		(*f)(p[i]);
 	}
